Add support for disabling mobile full-screen flag dropdown
When mobile useragent is detected, the flag dropdown is rendered as a
direct child of `document.body`. This might be undesirable for
use-cases when the telephone input must be strictly contained in a
certain DOM element.

Add `useMobileFullScreenDropdown` prop to `IntlTelInputApp` which specifies
whether to use the _fullscreen_ rendering for mobile useragents.
`useMobileFullScreenDropdown` defaults to `true` and is thus backwards
compatible.