Add findIndex implementation for IE 11
`findIndex` is not available in IE 11.

There is `core-js` polyfill available [1], however would like to avoid
using it as our app runs on top of other sites as a widget and we do not
want to define any kinds of globally available functions.

Instead use some which is implemented in IE 11.

[1]: https://github.com/zloirock/core-js/tree/v2#ecmascript-6-array